Simplified and updated documentation
